hirepath_fixed/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pypdf
python-docx
openai
